Migrate Pickup to draft
