+++ templates/alliance_management.html
@@ -72,7 +72,7 @@
 			{% endif %}	
 			{% endfor %}
 			<h2>Other options:</h2>
-			<a href="/pickUp/">Pick up free agent</a>
+			<a href="/draft/pickUp/">Pick up free agent</a>
 	</div>
 	<nav>
 	  <ul class="pager">
